+++ kernel/scheduling/userevent.c
@@ -30,7 +30,7 @@
 #include <ioset.h>
 
 typedef struct UserEvent {
-    size_t       initalValue;
+    unsigned int initalValue;
     unsigned int flags;
     atomic_int*  sync_address;
 } UserEvent_t;
@@ -65,7 +65,7 @@ UserEventDestroy(
 
 OsStatus_t
 UserEventCreate(
-    _In_  size_t       initialValue,
+    _In_  unsigned int initialValue,
     _In_  unsigned int flags,
     _Out_ UUId_t*      handleOut,
     _Out_ atomic_int** syncAddressOut)
